Refactor out getAppInsightsSupprotedLocation and use it in LogAnalyticsCreateStep
